_GNU_SOURCE

+++ util.c
@@ -548,7 +548,7 @@ error(char *format, ...) {
     }
 
     buffer[n] = '\0';
-    write(STDERR_FILENO, buffer, n);
+    write(STDERR_FILENO, buffer, (size_t)n);
 #if OS_UNIX
     fsync(STDERR_FILENO);
     fsync(STDOUT_FILENO);
